Increase timeout in integration tests
Sometimes the services that are brought up with docker-compose may take
a little while before they become available.
I'm adding a 3 seconds timeout to each of the steps that need to somehow
wait for some AJAX to load and a 30 seconds sleep time before the tests
start to kick in after bringing the cluster up with docker-compose.